Change to receive parameters (ID) in the URL for the API to return the ID data
Creating the /users/:id route

GetUserByID function in the controller

In index.js, I created a splitEndpoint calling the splint function to split my endpoint and separate it into an array, thus removing the space.

In the controller, I destructured my ID and performed a find of my user from the users array (mock) to search for the ID and register it.