Correcting HTTP to HTTP in parsedUrl (index.js)
Capturing the request body

Creating a new route to create a new user

Creating createUser in the Controller file:
In it, I created an event listener (request.on) to listen and collect all the information.

I also created an 'end' event listener to know when it was finished and receive the last chunk of the chunk.

I also created a constant within my request.on('end') to get the last registered user ID and then create a new user (newUser), passing the ID with the lastUserID and the name.

After the entire event listener process and receiving all the chunks, I pushed the array and then a response.send to return 200 and the registered user's data.