Adjusting the chunk name
Creating the bodyParser.js file

Removing the eventlistener section to centralize the bodyparser and avoid cluttering the code every time a POST, PUT, or PATCH method is called.

Creating a callback, since the bodyParser is receiving messages gradually, and I can only execute the handler after the 'end' is executed.

Inside my index.js, I created an if statement to execute my route handler and an array with the PUT, POST, and PATCH methods.
I used .includes in the array to check if the string with the methods is within my array.

Finally, I updated the userController, bringing the {body} destructor from my request and the logic for registering a new user.