Creating Edit and Delete Routes
In the controller, I created updateUser and deleteUser.

In the update, I destructured my ID because it's in my request.params. I also got the name from request.body.

I updated my array from const to let users in the controller call.

In the update, I used the find method to perform a "search" to check if any user has the same ID as the one I'm receiving in my URL. After that, I created a map to check if the user.id I'm currently interacting with is the same as the ID I'm editing. I return an object with the current user's information and overwrite the name.

In the delete, I perform the same procedure as the ID and get my users who will receive the filter method to remove the users who have the same ID as the one I'm sending in my URL.